Move Zuplo env file
